requirements.txt
datasets
transformers
sentence-transformers
faiss-cpu
streamlit
pandas
numpy
ragas
langchain
langchain-community
torch
ollama